Change interface type to uint32_t from size_t

+++ plugins/core/widgets/src/Madgine/widgets/tabbar.cpp
@@ -77,12 +77,12 @@ namespace Widgets {
         return WidgetClass::TABBAR;
     }
 
-    size_t TabBar::tabCount() const
+    uint32_t TabBar::tabCount() const
     {
         return mTabNames.size();
     }
 
-    void TabBar::setTabCount(size_t count)
+    void TabBar::setTabCount(uint32_t count)
     {
         mTabNames.resize(count);
         applyGeometry();
